+++ Gemfile
@@ -16,7 +16,7 @@ gem "tzinfo-data", platforms: %i[ windows jruby ]
 gem "geocoder", "~> 1.6"
 
 # Use Rack CORS for handling Cross-Origin Resource Sharing (CORS), making cross-origin Ajax possible
-# gem "rack-cors"
+gem "rack-cors"
 
 group :development, :test do
   # See https://guides.rubyonrails.org/debugging_rails_applications.html#debugging-with-the-debug-gem
