+++ test/controllers/locations_controller_test.rb
@@ -7,31 +7,31 @@ class LocationsControllerTest < ActionDispatch::IntegrationTest
   end
 
   test "should get index" do
-    get locations_url, as: :json
+    get api_v1_locations_url, as: :json
     assert_response :success
   end
 
   test "should create location" do
     assert_difference("Location.count") do
-      post locations_url, params: { location: { address: @location.address } }, as: :json
+      post api_v1_locations_url, params: { location: { address: @location.address } }, as: :json
     end
 
     assert_response :created
   end
 
   test "should show location" do
-    get location_url(@location), as: :json
+    get api_v1_location_url(@location), as: :json
     assert_response :success
   end
 
   test "should update location" do
-    patch location_url(@location), params: { location: { address: @location.address } }, as: :json
+    patch api_v1_location_url(@location), params: { location: { address: @location.address } }, as: :json
     assert_response :success
   end
 
   test "should destroy location" do
     assert_difference("Location.count", -1) do
-      delete location_url(@location), as: :json
+      delete api_v1_location_url(@location), as: :json
     end
 
     assert_response :no_content
@@ -40,7 +40,7 @@ class LocationsControllerTest < ActionDispatch::IntegrationTest
   test "should save invalid address with error code 404" do
     @invalid_address = locations(:invalid_address)
     assert_difference("Location.count") do
-      post locations_url, params: { location: { address: @invalid_address.address } }, as: :json
+      post api_v1_locations_url, params: { location: { address: @invalid_address.address } }, as: :json
     end
 
     assert_equal 404, Location.last.geocode_error
@@ -49,7 +49,7 @@ class LocationsControllerTest < ActionDispatch::IntegrationTest
   test "should save only unique invalid addresses with error code 404" do
     @invalid_address = locations(:invalid_address)
     assert_difference("Location.count") do
-      post locations_url, params: { location: { address: @invalid_address.address } }, as: :json
+      post api_v1_locations_url, params: { location: { address: @invalid_address.address } }, as: :json
     end
 
     assert_equal 404, Location.last.geocode_error
